✅ added more tests
